[hotfix][operator] Deduplicate code by introducing StreamInputProcessor interface
