+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/TwoInputStreamTask.java
@@ -33,8 +33,6 @@ import java.util.Collection;
 @Internal
 public class TwoInputStreamTask<IN1, IN2, OUT> extends AbstractTwoInputStreamTask<IN1, IN2, OUT> {
 
-	private StreamTwoInputProcessor<IN1, IN2> inputProcessor;
-
 	public TwoInputStreamTask(Environment env) {
 		super(env);
 	}
@@ -61,19 +59,5 @@ public class TwoInputStreamTask<IN1, IN2, OUT> extends AbstractTwoInputStreamTas
 			input2WatermarkGauge,
 			getTaskNameWithSubtaskAndId(),
 			operatorChain);
-	}
-
-	@Override
-	protected void performDefaultAction(ActionContext context) throws Exception {
-		if (!inputProcessor.processInput()) {
-			context.allActionsCompleted();
-		}
-	}
-
-	@Override
-	protected void cleanup() throws Exception {
-		if (inputProcessor != null) {
-			inputProcessor.cleanup();
-		}
 	}
 }
